// frontend/src/common/api/generated/models/socketClientServerAddSensor.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * dashboard API
 * OpenAPI spec version: 1.0.0
 */
import type { MessageType } from './messageType';
import type { OffsetDateTime } from './offsetDateTime';
import type { OriginType } from './originType';
import type { SocketClientServerAddSensorOriginIdentifier } from './socketClientServerAddSensorOriginIdentifier';
import type { SocketClientServerAddSensorContent } from './socketClientServerAddSensorContent';

export interface SocketClientServerAddSensor {
  type: MessageType;
  date: OffsetDateTime;
  origin: OriginType;
  originIdentifier?: SocketClientServerAddSensorOriginIdentifier;
  content?: SocketClientServerAddSensorContent;
}